style(ms2/change-email): better feedback
