+++ src/management-system-v2/app/change-email/page.tsx
@@ -1,12 +1,9 @@
 import { getCurrentUser } from '@/components/auth';
-import Content from '@/components/content';
-import { requestEmailChange } from '@/lib/change-email/server-actions';
 import { getTokenHash, notExpired } from '@/lib/change-email/utils';
 import { getVerificationToken } from '@/lib/data/legacy/verification-tokens';
-import { Button, Card, Space } from 'antd';
 import { redirect } from 'next/navigation';
 import { z } from 'zod';
-import ConfirmationButtons from './confirmation-buttons';
+import ChangeEmailCard from './change-email-card';
 
 const searchParamsScema = z.object({ email: z.string().email(), token: z.string() });
 
@@ -17,7 +14,8 @@ export default async function ChangeEmailPage({ searchParams }: { searchParams: 
 
   const { session } = await getCurrentUser();
   const userId = session?.user.id;
-  if (!userId) redirect('/');
+  if (!userId || session.user.guest) redirect('/');
+  const previousEmail = session.user.email;
 
   const verificationToken = getVerificationToken({
     identifier: email,
@@ -32,11 +30,5 @@ export default async function ChangeEmailPage({ searchParams }: { searchParams: 
   )
     redirect('/');
 
-  return (
-    <Content title="Change Email">
-      <Card title="Confirm Email change">
-        <ConfirmationButtons />
-      </Card>
-    </Content>
-  );
+  return <ChangeEmailCard previousEmail={previousEmail} newEmail={email} />;
 }
